Moved logic to determine if a user is allowed to edit a project to the application controller. I'm not a fan of this but I couldn't get the resource owner in a controller concern. The motivation of this was to be able to resue it on the controllers to add and remove layers to a project.

+++ app/controllers/api/v1/raster_layer_projects_controller.rb
@@ -43,8 +43,12 @@ class Api::V1::RasterLayerProjectsController < ApplicationController
 
   def destroy
     projectlayer = RasterLayerProject.find(params[:id])
-    projectlayer.destroy
-    head 204
+    if mine(projectlayer.project) == true
+      projectlayer.destroy
+      head 204
+    else
+      head 401
+    end
   end
 
   private
